+++ kani-compiler/src/kani_middle/transform/body.rs
@@ -2,6 +2,8 @@
 // SPDX-License-Identifier: Apache-2.0 OR MIT
 //
 //! Utility functions that allow us to modify a function body.
+//!
+//! TODO: We should not need this code anymore now that https://github.com/rust-lang/rust/pull/138536 is merged
 
 use crate::kani_middle::kani_functions::KaniHook;
 use crate::kani_queries::QueryDb;
@@ -54,12 +56,10 @@ impl MutableBody {
         &self.locals
     }
 
-    #[allow(dead_code)]
     pub fn arg_count(&self) -> usize {
         self.arg_count
     }
 
-    #[allow(dead_code)]
     pub fn var_debug_info(&self) -> &Vec<VarDebugInfo> {
         &self.var_debug_info
     }
@@ -328,7 +328,6 @@ impl MutableBody {
     /// `InsertPosition` is `InsertPosition::Before`, `source` will point to the same instruction as
     /// before. If `InsertPosition` is `InsertPosition::After`, `source` will point to the
     /// terminator of the newly inserted basic block.
-    #[allow(dead_code)]
     pub fn insert_bb(
         &mut self,
         mut bb: BasicBlock,